chore: ping preview supabase
